requirements.txt
pandas>=1.5.0
pyarrow>=10.0.0
numpy>=1.23.0
scikit-learn>=1.2.0
pyyaml>=6.0
streamlit>=1.28.0
plotly>=5.18.0
streamlit-option-menu>=0.3.6
streamlit-authenticator>=0.3.1
bcrypt>=4.0.0
